pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apcore-a2a"
version = "0.1.0"
description = "Automatic A2A Protocol Adapter for apcore Module Registry"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "aipartnerup", email = "team@aipartnerup.com" },
]
keywords = ["apcore", "a2a", "agent", "ai", "server", "client", "protocol", "jwt", "streaming", "webhook"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "apcore>=0.7.0",
    "a2a-sdk>=0.3.20",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "PyJWT>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "httpx>=0.27.0",
    "apdev[dev]>=0.2.0",
]

[project.scripts]
apcore-a2a = "apcore_a2a.__main__:main"

[project.urls]
Homepage = "https://github.com/aipartnerup"
Repository = "https://github.com/aipartnerup/apcore-a2a-python"
Documentation = "https://github.com/aipartnerup/apcore-a2a-python#readme"
Issues = "https://github.com/aipartnerup/apcore-a2a-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/apcore_a2a"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.coverage.run]
source = ["src/apcore_a2a"]
omit = ["src/apcore_a2a/__main__.py"]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.apdev]
base_package = "apcore_a2a"
src_dir = "src"
